Added namespace to SimpleHydratorTest
