+++ tests/Assets/Testing/AncestorHydrateableClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Testing;
+
+class AncestorHydrateableClass
+{
+    private $name;
+
+    public function getName()
+    {
+        return $this->name;
+    }
+}
