+++ tests/Assets/Testing/ChildHydrateableClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Testing;
+
+class ChildHydrateableClass extends AncestorHydrateableClass
+{
+    private $description;
+
+    public function getDescription()
+    {
+        return $this->description;
+    }
+}
