+++ tests/Feature/Testing/SimpleHydratorTest.php
@@ -1,17 +1,21 @@
 <?php
 
+namespace LaravelDoctrineTest\ORM\Feature\Testing;
+
 use LaravelDoctrine\ORM\Testing\SimpleHydrator;
+use LaravelDoctrineTest\ORM\Assets\Testing\AncestorHydrateableClass;
+use LaravelDoctrineTest\ORM\Assets\Testing\ChildHydrateableClass;
 use PHPUnit\Framework\TestCase;
 
 class SimpleHydratorTest extends TestCase
 {
     public function test_can_hydrate_class()
     {
-        $entity = SimpleHydrator::hydrate(BaseHydrateableClass::class, [
+        $entity = SimpleHydrator::hydrate(AncestorHydrateableClass::class, [
             'name' => 'Patrick',
         ]);
 
-        $this->assertInstanceOf(BaseHydrateableClass::class, $entity);
+        $this->assertInstanceOf(AncestorHydrateableClass::class, $entity);
         $this->assertEquals('Patrick', $entity->getName());
     }
 
@@ -25,25 +29,5 @@ class SimpleHydratorTest extends TestCase
         $this->assertInstanceOf(ChildHydrateableClass::class, $entity);
         $this->assertEquals('Patrick', $entity->getName());
         $this->assertEquals('Hello World', $entity->getDescription());
-    }
-}
-
-class ChildHydrateableClass extends BaseHydrateableClass
-{
-    private $description;
-
-    public function getDescription()
-    {
-        return $this->description;
-    }
-}
-
-class BaseHydrateableClass
-{
-    private $name;
-
-    public function getName()
-    {
-        return $this->name;
     }
 }
